La til søke funksjon på hjemmesiden, og fjernet den dårlige vi hadde før.
