+++ app/components/search/results.tsx
@@ -6,30 +6,50 @@ import Arrangement from "./arrangement"
 import Link from "next/link";
 
 export default function Results() {
-    const [event, setEvent] = useState([]);
+    const [searchTerm, setSearchTerm] = useState<string>('');
+    const [events, setEvents] = useState<any[]>([]);
 
+    // Function to handle the search input change
+    const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setSearchTerm(event.target.value);
+    };
+
+    // Function to fetch events based on the search term
+    const fetchEvents = async () => {
+        try {
+            const response = await fetch(searchTerm ? `/api/search?name=${encodeURIComponent(searchTerm)}` : '/api/table');
+            const data = await response.json();
+            setEvents(data);
+        } catch (error) {
+            console.error('Error fetching events:', error);
+        }
+    };
+
+    // Call fetchEvents whenever the searchTerm changes
     useEffect(() => {
-        const fetchData = async () => {
-        const response = await fetch('/api/table');
-        const data = await response.json();
-        setEvent(data);
-        };
-
-        fetchData();
-    }, []);
+        fetchEvents();
+    }, [searchTerm]);
 
     return (
-        <div className="mainBox flex-col basis-1/2 min-h-1/2">
-            {event && event.length && event.map((event:any, index) => (
-                <ul key={index}>
-                    <Link href={{
-                        pathname: '/infoPage',
-                        query: { id: event.id, }
-                    }}>
-                        <Arrangement title={event.name} desc={event.description} pic={event.picture_normal}/>
-                    </Link>
-                </ul>
-            ))}
+        <div>
+            <input 
+            type="text"
+            placeholder="Search for events..."
+            value={searchTerm}
+            onChange={handleSearchChange}
+            />
+            <div className="mainBox flex-col basis-1/2 min-h-1/2">
+                {events && events.length > 0 ? events.map((event:any, index) => (
+                    <ul key={index}>
+                        <Link href={{
+                            pathname: '/infoPage',
+                            query: { id: event.id, }
+                        }}>
+                            <Arrangement title={event.title} desc={event.name} pic={event.picture_normal}/>
+                        </Link>
+                    </ul>
+                )) : <p>Ingen resultat</p>}
+            </div>
         </div>
     )
 }
